:wrench: (backend) updated analysis for saas metrics
