+++ backend/src/core/services/worker/cloud_tasks.py
@@ -1,5 +1,6 @@
-from google.cloud import tasks_v2
 import json
+
+from google.cloud import tasks_v2
 
 
 def create_task_payload(worker_url, payload):
